Reversed Chains
Description
Given a singly linked list, reverse it so that the tail becomes the new head, and the head is the new tail.

This problem uses a solution framework. Implement the missing function in the framework to complete the program.

Input:
On the first line of the input is a number, which is the number of integers to follow.

On the second line is the list of integers, separated by one or more space(s), followed by a trailing newline.

Output:
The integers in the reversed order. An optional trailing newline is accepted.