Concatenable Intervals
Description
Design and implement an iterator that holds a series of intervals and iterates over the integers that are covered by the intervals. Additional intervals can be "concatenated" (i.e. the union of the current and the new interval) by calling a method of the iterator.

Input
On the first line of the input is a number indicating how many lines of instructions to follow.

On each of the following lines is an instruction to the iterator, either to append an additional interval, or to iterate one step over the exisiting intervals.

If the line starts with a "+" sign, and followed by two integers seperated by a comma, concatenate the interval noted by the two integers to the iterator; if the line contains only a "-" sign, iterate one step forward and output the next integer. If the iterator is an empty set, return INT_MIN.

All the intervals are finite, left-closed and right-open, and both ends are integers (can be posistive or negative).

Output
The integers during iteration, one on each line.